Fix bug
There was a bug in how the friends list was massaged. The friends
list is a comma separated string. When a friend is added, the
list is appended with the id of the new friend. However, if that
id contained a number previously found in the list, the id
would get corrupted. An example:
original: 2,3,8
appended: 2,3,8,21
final: 2,3,8,1

The fix is to treat the string split on commas as a set.